Factor out a DiagnosticMessageGenerator class from DiagnosticReporterFactory.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=282615199
